[OPP-1321] flytte hotjar-trigger
flyttet slik at den er gjelde i hele applikasjonen, og ikke bare på startsiden. Gjøres slik at vi også fanger opp folk som kommer direkte fra gosys
